requirements.txt
# External APIs
google-api-python-client==1.7.11
google-auth==1.10.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
Pillow>=7.0.0
ffmpeg-python>=0.2.0